+++ quik/src/main/java/com/example/quik/adapter/messages/QLQuote.java
@@ -4,14 +4,12 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigDecimal;
-
 @Getter
 @Setter
 public final class QLQuote {
 
     @JsonProperty("p")
-    private BigDecimal price;
+    private double price;
 
     @JsonProperty("q")
     private int size;
